+++ sdk/websites/Azure.ResourceManager.AppService/CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ### Features Added
 
+- Enable the new model serialization by using the System.ClientModel, refer this [document](https://aka.ms/azsdk/net/mrw) for more details.
+
 ### Breaking Changes
 
 ### Bugs Fixed
+
+-Add `200` to the response code of the `SyncFunctionTriggers`
 
 ### Other Changes
 
@@ -59,7 +63,7 @@
   `SiteAuthSettings.ClientSecretCertificateThumbprint` has been hidden but is still available.
 - Added `PublicCertificateData.ThumbprintString` to return the hexadecimal string representation of the SHA-1 hash of the certificate.
   `PublicCertificateData.Thumbprint` has been hidden but is still available.
-  
+
 ### Other Changes
 
 - Upgraded dependent Azure.Core to 1.32.0.